+++ internal/http/http.go
@@ -37,6 +37,7 @@ func NewHandler(pool *pgxpool.Pool, a *auth.Service, blobs *media.Store) (http.H
 	registerReviewRoutes(mux, pool, pages, fragments, time.Now)
 	registerMediaRoutes(mux, pool, blobs)
 	registerImportRoutes(mux, pool, pages, blobs, time.Now)
+	registerAIImportRoutes(mux, pool, pages)
 	return securityHeaders(a.Middleware(mux)), nil
 }